+++ src/store/actions.js
@@ -18,5 +18,5 @@ export * from './notices/selectors'
 export * from './profiles/selectors'
 export * from './opportunities/selectors'
 export * from './modals/actions'
-export * from './applicants/actions'
-export * from './applicants/selectors'
+export * from './offers/actions'
+export * from './offers/selectors'
