--- src/store/applicants/selectors.js
@@ -1,8 +1,0 @@
-import { createSelector } from 'reselect'
-
-export const applicantsSelector = () => (state) => state.applicants
-
-export const jobsApplicantsSelector = (jobId) => createSelector(
-  applicantsSelector(),
-  (applicants) => applicants.byJobKey[`job-${jobId}`]
-)
