+++ src/store/index.js
@@ -7,7 +7,7 @@ import { createLogger } from 'redux-logger'
 import * as thunkMiddleware from 'redux-thunk'
 import web3Factory from 'src/web3'
 import { initFrontend, initBackend } from './bootstrap'
-import { login, landing, ethereum, daos, tokens, wallet, balances, createAccount, boards, jobs, modals, user, applicants } from './reducers'
+import { login, landing, ethereum, daos, tokens, wallet, balances, createAccount, boards, jobs, modals, user, offers } from './reducers'
 
 export * from './actions'
 
@@ -39,7 +39,7 @@ export default (initialState = {}) => {
     createAccount,
     modals,
     user: user(),
-    applicants,
+    offers,
   })
 
   // Here you can recover state sent from the backend
