+++ src/store/jobs/reducer.js
@@ -1,6 +1,5 @@
 import {
   JOBS_CLEAR, JOBS_SAVE, JOBS_FILTER, JOBS_CLIENT_SAVE, JOBS_WORKER_SAVE,
-  GET_JOBS_OFFERS_REQUEST, GET_JOBS_OFFERS_FAILURE, GET_JOBS_OFFERS_SUCCESS,
 } from './actions'
 
 export const initialState = {
@@ -11,11 +10,6 @@ export const initialState = {
   byKey: {},
   clientByKey: {},
   workerByKey: {},
-  offers: {
-    loading: false,
-    offers: null,
-    err: null,
-  },
 }
 
 const mutations = {
@@ -60,11 +54,6 @@ const mutations = {
       filtered: jobsList,
     }
   },
-
-  [GET_JOBS_OFFERS_REQUEST]: (state) => ({ ...state, offers: { ...state.offers, loading: true } }),
-  [GET_JOBS_OFFERS_SUCCESS]: (state, { offers }) => ({ ...state, offers: { ...state.offers, loading: false, offers } }),
-  [GET_JOBS_OFFERS_FAILURE]: (state, { err }) => ({ ...state, offers: { ...state.offers, loading: false, err } }),
-
 }
 
 export default (state = initialState, { type, ...other }) => {
