+++ src/store/offers/selectors.js
@@ -1,0 +1,8 @@
+import { createSelector } from 'reselect'
+
+export const offersSelector = () => (state) => state.offers
+
+export const jobsOffersSelector = (jobId) => createSelector(
+  offersSelector(),
+  (offers) => offers.byJobKey[`job-${jobId}`]
+)
