+++ src/store/reducers.js
@@ -10,4 +10,4 @@ export { default as boards } from './boards/reducer'
 export { default as jobs } from './jobs/reducer'
 export { default as modals } from './modals/reducer'
 export { default as user } from './user/reducer'
-export { default as applicants } from './applicants/reducer'
+export { default as offers } from './offers/reducer'
